gk: Initialize V from message

+++ common/types/src/lib.rs
@@ -246,8 +246,8 @@ pub mod messaging {
     #[derive(Encode, Decode, Clone, Debug)]
     pub struct SettleInfo {
         pub pubkey: WorkerPublicKey,
-        pub v: u32,
-        pub payout: u32,
+        pub v: u64,
+        pub payout: u64,
     }
 }
